+++ src/repositories/post/PostRepositoryImpl.ts
@@ -1,87 +1,51 @@
 import { PostRepository } from './PostReppository'
-import { ImageTag, Post, PostImage } from 'types/entities/Post'
-import {
-  ImageTagReturnType,
-  PostImageReturnType,
-  PostReturnType,
-  postMapper,
-} from './PostMapper'
+import { Post } from 'types/entities/Post'
+import { postMapper } from './PostMapper'
+import { Tables, supabase } from 'plugins/supabse'
 
 class PostRepositoryImpl implements PostRepository {
+  private selectQueryBuilder = () => {
+    return supabase.from('posts').select(`
+        *,
+        post_images (
+            *,
+            image_tags (
+                *
+            )
+        ),
+        address: addresses (
+            *
+        )
+        `)
+  }
   public findAll = async (): Promise<Post[]> => {
-    const res = await postMapper.findAll()
-    return res?.map(convert) ?? []
+    const { data } = await this.selectQueryBuilder().returns<PostReturnType[]>()
+    return postMapper.findAll(data)
   }
-  public findById = async (id: string): Promise<Post | undefined> => {
-    const res = await postMapper.findById(id)
-    if (!res) return undefined
-    return convert(res)
+  public findById = async (postId: string): Promise<Post | undefined> => {
+    const { data } = await this.selectQueryBuilder()
+      .eq('post_id', postId)
+      .single()
+    return postMapper.findById(data as PostReturnType | null)
   }
+
   public save = async (post: Post): Promise<void> => {
-    await postMapper.save({
-      user_id: post.userId,
-      title: post.title,
-      comment: post.comment,
-      posted_at: post.postedAt,
-      post_id: post.postId,
-    })
-    await postMapper.saveAddress({
-      address_id: post.address.addressId,
-      latitude: post.address.latitude,
-      longitude: post.address.longitude,
-      post_id: post.postId,
-    })
-    await this.saveImages(post.postImages, post.postId)
-    return
-  }
-  private saveImages = async (postImages: PostImage[], postId: string) => {
-    postImages.forEach(async (postImage) => {
-      await postMapper.saveImage({
-        post_image_id: postImage.postImageId,
-        image_path: postImage.imagePath,
-        post_id: postId,
-      })
-      postImage.imageTags.forEach(async (imageTag) => {
-        await postMapper.saveImageTag({
-          name: imageTag.name,
-          post_image_id: postImage.postImageId,
-          tag_id: imageTag.tagId,
-        })
-      })
-    })
+    await supabase.from('posts').insert(postMapper.save(post))
+    await supabase.from('addresses').insert(postMapper.saveAddress(post))
+    await supabase.from('post_images').insert(postMapper.saveImage(post))
+    await supabase.from('image_tags').insert(postMapper.saveImageTag(post))
     return
   }
 }
 
 export const postRepository = new PostRepositoryImpl()
 
-const convert = (res: PostReturnType): Post => {
-  return {
-    postId: res.post_id,
-    userId: res.user_id,
-    title: res.title,
-    comment: res.comment ?? undefined,
-    postedAt: res.posted_at,
-    address: {
-      addressId: res.address.address_id,
-      latitude: res.address.latitude,
-      longitude: res.address.longitude,
-    },
-    postImages: res.post_images.map(convertPostImage),
-  }
+export type PostReturnType = Tables['posts']['Row'] & {
+  post_images: PostImageReturnType[]
+  address: AddressReturnType
 }
-
-const convertPostImage = (res: PostImageReturnType): PostImage => {
-  return {
-    postImageId: res.post_image_id,
-    imagePath: res.image_path,
-    imageTags: res.image_tags.map(convertImageTag),
-  }
+export type PostImageReturnType = Tables['post_images']['Row'] & {
+  image_tags: ImageTagReturnType[]
 }
-
-const convertImageTag = (res: ImageTagReturnType): ImageTag => {
-  return {
-    name: res.name,
-    tagId: res.tag_id,
-  }
-}
+export type ImageTagReturnType = Tables['image_tags']['Row']
+export type AddressReturnType = Tables['addresses']['Row']
